Refactor: Move the assign functions toe the assign package
